Version 3!
I'm still working on the input system, but this is far better than on the command line (in previous versions).
I've added a sudoku checker and something that sees how many solutions are possible. If there are 5000 or more solutions, it just says '5000 or more'. This is because there are over 6.5 sextillion sudoku boards, and if you inputted a blank grid, the program may take a couple of years. I will work on a feature where you can adjust the limit, but 5000 will remain the default.