I am asking people to test this
This folder is purely for the people I am asking to test this and try to break the program. If you do find a bug or think of a new feature I'm capable of coding, please let me know. You can open it as an issue (even if it is a new feature - its absence is an issue).