Does this actually work??
I have high hopes for this compilation.
If everything goes to plan, you will ONLY NEED THIS FILE TO RUN THE PROGRAM!!!!
Pyinstaller stopped working for me because of a bug in the new version, so I searched for alternatives. The one I used for this is called Nuitka.
In hindsight, I could have actually just adjusted my command and code a bit and used pyinstaller for this result.
But, I'll run some tests and hope this works.